migrating Openstack, cinder and swift to the collections mixin
