+++ app/models/manageiq/providers/openstack/cloud_manager/refresh_parser.rb
@@ -3,7 +3,7 @@
 module ManageIQ::Providers
   class Openstack::CloudManager::RefreshParser < ManageIQ::Providers::CloudManager::RefreshParser
     include Vmdb::Logging
-    include ManageIQ::Providers::Openstack::RefreshParserCommon::HelperMethods
+    include CollectionsParserMixin
     include ManageIQ::Providers::Openstack::RefreshParserCommon::Images
     include ManageIQ::Providers::Openstack::RefreshParserCommon::Objects
     include ManageIQ::Providers::Openstack::RefreshParserCommon::Flavors
